+++ src/main/java/ac/boar/anticheat/packets/input/AuthInputPackets.java
@@ -1,6 +1,8 @@
 package ac.boar.anticheat.packets.input;
 
+import ac.boar.anticheat.check.impl.timer.Timer;
 import ac.boar.anticheat.data.input.PredictionData;
+import ac.boar.anticheat.data.input.TickData;
 import ac.boar.anticheat.data.input.VelocityData;
 import ac.boar.anticheat.packets.input.legacy.LegacyAuthInputPackets;
 import ac.boar.anticheat.player.BoarPlayer;
@@ -31,11 +33,36 @@ public class AuthInputPackets implements PacketListener {
             return;
         }
 
+        // TODO: ugggggh, current timer logic is a bit broken.
+        // -------------------------------------------------------------------------
+        Timer timer = (Timer) player.getCheckHolder().get(Timer.class);
+        if (player.tick == Long.MIN_VALUE) {
+            if (timer == null) {
+                player.tick = Math.max(0, packet.getTick()) - 1;
+            }
+        }
+        if (timer != null) {
+            if (timer.tick(packet.getTick())) {
+                return;
+            }
+        } else {
+            player.tick++;
+            if (packet.getTick() != player.tick || packet.getTick() < 0) {
+                player.kick("Invalid tick id=" + packet.getTick());
+                return;
+            }
+        }
+        // -------------------------------------------------------------------------
+
         player.breakingValidator.handle(packet);
         player.tick();
 
         LegacyAuthInputPackets.processAuthInput(player, packet, true);
         LegacyAuthInputPackets.updateUnvalidatedPosition(player, packet);
+
+        if (player.vehicleData != null) { // TODO: Vehicle prediction.
+            return;
+        }
 
         if (player.isMovementExempted()) {
             player.setPos(player.unvalidatedPosition);
@@ -55,12 +82,13 @@ public class AuthInputPackets implements PacketListener {
             // This is fine, we only need tick end and use before and after to calculate ground.
             player.predictionResult = new PredictionData(Vec3.ZERO, player.velocity.y < 0 && player.getInputData().contains(PlayerAuthInputData.VERTICAL_COLLISION) ? new Vec3(0, 1, 0) : Vec3.ZERO, player.unvalidatedTickEnd);
             player.velocity = player.unvalidatedTickEnd.clone();
+
+            player.bestPossibility = Vector.NONE;
         } else {
             new PredictionRunner(player).run();
         }
         this.processQueuedTeleports(player, packet);
-
-        player.getTeleportUtil().cachePosition(player.tick, player.position.add(0, player.getYOffset(), 0).toVector3f());
+        LegacyAuthInputPackets.doPostPrediction(player, packet);
     }
 
     private void processQueuedTeleports(final BoarPlayer player, final PlayerAuthInputPacket packet) {
@@ -128,6 +156,50 @@ public class AuthInputPackets implements PacketListener {
         }
 
         return true;
+    }
+
+    // Wouldn't it be nice to provide us a way to know when player accept rewind mojang :(
+    // There will be edge cases where player lag right after accepting the stack id responsible for rewind... maaaaaaaybe
+    // we could check for current offset and if it's close then player might possibility HAVEN'T received the rewind yet?
+    // Just ignore the edge cases for now.
+    private void processRewind(final BoarPlayer player, final TeleportCache.Rewind rewind, final PlayerAuthInputPacket packet) {
+        if (player.isMovementExempted()) { // Fully exempted from rewind teleport.
+            return;
+        }
+
+        player.onGround = rewind.isOnGround();
+        player.velocity = rewind.getTickEnd();
+        player.setPos(rewind.getPosition().subtract(0, player.getYOffset(), 0));
+        player.prevUnvalidatedPosition = player.unvalidatedPosition = player.position.clone();
+
+        player.getTeleportUtil().cachePosition(rewind.getTick(), rewind.getPosition().toVector3f());
+
+        // Keep running prediction until we catch up with the player current tick.
+        long currentTick = rewind.getTick();
+        while (currentTick != player.tick) {
+            if (currentTick != rewind.getTick() && player.position.distanceTo(player.unvalidatedPosition) > player.getMaxOffset()) {
+                player.unvalidatedPosition = player.position.clone();
+            }
+
+            currentTick++;
+
+            if (currentTick == player.tick) {
+                LegacyAuthInputPackets.processAuthInput(player, packet, true);
+                LegacyAuthInputPackets.updateUnvalidatedPosition(player, packet);
+            } else if (player.getTeleportUtil().getAuthInputHistory().containsKey(currentTick)) {
+                final TickData data = player.getTeleportUtil().getAuthInputHistory().get(currentTick);
+                LegacyAuthInputPackets.processAuthInput(player, data.packet(), false);
+                LegacyAuthInputPackets.updateUnvalidatedPosition(player, packet);
+
+                // Reverted back to the old flags.
+                player.getFlagTracker().set(data.flags(), false);
+            } else {
+                throw new RuntimeException("Failed find auth input history for rewind.");
+            }
+
+            new PredictionRunner(player).run();
+            // player.getTeleportUtil().cachePosition(currentTick, player.position.add(0, player.getYOffset(), 0).toVector3f());
+        }
     }
 
     @Override
